Fixing the repository.
Git has version control, we don't need to make new folders for different
versions. Simply update the files and do a push. Thanks!

Removed library and binaries, because that will cause headaches. We only
need to sync assets, project settings, and potentially the prefs files.

+++ Assets/Scripts/StatCollectionClass.cs
@@ -5,7 +5,7 @@ public class StatCollectionClass : MonoBehaviour {
 	
 	public int health;
 	
-	public int mana;
+	public float mana;
 		
 	public int strength;
